Moving tripleo utils to kernel tools.
