+++ kernel/tools/paths.py
