+++ kernel/tools/rng.py
