+++ kernel/tools/strings.py
