+++ tests/kernel/unit/tools/test_urls.py
@@ -16,14 +16,14 @@
 from unittest import TestCase
 from unittest.mock import Mock, patch
 
-from tripleo.utils.urls import URL
+from kernel.tools.urls import URL
 
 
 class TestURL(TestCase):
     """Tests for :class:`URL`.
     """
 
-    @patch('tripleo.utils.urls.is_url')
+    @patch('kernel.tools.urls.is_url')
     def test_new_on_valid_url(self, check_mock: Mock):
         """Checks that if the string is a URL, the string is created.
         """
@@ -35,7 +35,7 @@ class TestURL(TestCase):
 
         check_mock.assert_called_once_with(string)
 
-    @patch('tripleo.utils.urls.is_url')
+    @patch('kernel.tools.urls.is_url')
     def test_error_on_invalid_url(self, check_mock: Mock):
         """Checks that if the string is not a URL, an error is thrown.
         """
@@ -48,7 +48,7 @@ class TestURL(TestCase):
 
         check_mock.assert_called_once_with(string)
 
-    @patch('tripleo.utils.urls.is_url')
+    @patch('kernel.tools.urls.is_url')
     def test_trims_string(self, check_mock: Mock):
         """Checks that before verifying the string, it is trimmed first.
         """
