+++ tripleo/insights/git.py
@@ -20,6 +20,7 @@ from typing import Iterable, Optional, Sequence
 from cached_property import cached_property
 from overrides import overrides
 
+from kernel.tools.fs import Dir
 from kernel.tools.git import Git, GitError
 from kernel.tools.git import Repository as GitRepo
 from kernel.tools.git.gitpython import GitPython
@@ -27,12 +28,11 @@ from kernel.tools.git.utils import get_repository_fullname
 from kernel.tools.github import GitHub, GitHubError
 from kernel.tools.github import Repository as GitHubRepo
 from kernel.tools.github.pygithub import PyGitHub
+from kernel.tools.paths import resolve_home
+from kernel.tools.rng import get_new_uuid
+from kernel.tools.urls import URL, is_git, is_github
+from kernel.tools.yaml import YAML, StandardYAMLParser, YAMLError, YAMLParser
 from tripleo.insights.exceptions import DownloadError, InvalidURL
-from tripleo.utils.fs import Dir
-from tripleo.utils.paths import resolve_home
-from tripleo.utils.rng import get_new_uuid
-from tripleo.utils.urls import URL, is_git, is_github
-from tripleo.utils.yaml import YAML, StandardYAMLParser, YAMLError, YAMLParser
 
 LOG = logging.getLogger(__name__)
